refactor(coral): Move Login into features.
