+++ coral/src/app/features/login/Login.tsx
@@ -1,0 +1,12 @@
+import { LoginForm } from "src/app/features/login/components/LoginForm";
+
+const Login = () => {
+  return (
+    <>
+      <h1>Login page</h1>
+      <LoginForm />
+    </>
+  );
+};
+
+export default Login;
