--- coral/src/app/features/login/index.ts
@@ -1,4 +1,0 @@
-import { LoginForm } from "src/app/features/login/LoginForm";
-import { useLoginUser } from "src/app/features/login/useLoginUser";
-
-export { LoginForm, useLoginUser };
