--- coral/src/app/pages/Login.tsx
@@ -1,12 +1,0 @@
-import { LoginForm } from "src/app/features/login";
-
-const LoginPage = () => {
-  return (
-    <>
-      <h1>Login page</h1>
-      <LoginForm />
-    </>
-  );
-};
-
-export default LoginPage;
