+++ coral/src/app/router.tsx
@@ -1,17 +1,12 @@
 import { createBrowserRouter, Navigate, RouteObject } from "react-router-dom";
 import HomePage from "src/app/pages";
 import Topics from "src/app/pages/topics";
-import Login from "src/app/pages/Login";
 import { getRouterBasename } from "src/config";
 
 const routes: Array<RouteObject> = [
   {
     path: "/",
     element: <HomePage />,
-  },
-  {
-    path: "/login",
-    element: <Login />,
   },
   {
     path: "/topics",
